Remove paymentParams prop from PaymentDialog
